Planner and Controller moved under base
