+++ morphac/CMakeLists.txt
@@ -16,7 +16,7 @@ add_subdirectory(controllers)
 add_subdirectory(environment)
 add_subdirectory(math)
 add_subdirectory(mechanics)
-#add_subdirectory(planners)
+add_subdirectory(planners)
 add_subdirectory(robot)
 add_subdirectory(simulation)
 add_subdirectory(utils)
