+++ morphac/controllers/base/CMakeLists.txt
@@ -7,7 +7,7 @@ set(CONTROLLERS_BASE_INCLUDE_DIR ${CONTROLLERS_BASE_DIR}/include)
 set(CONTROLLERS_BASE_TEST_DIR ${CONTROLLERS_BASE_DIR}/test)
 set(CONTROLLERS_BASE_PYTHON_DIR ${CONTROLLERS_BASE_DIR}/python)
 
-set(CONTROLLERS_BASE_BINARY_DIR ${CONTROLLERS_BINARY_DIR}/controllers)
+set(CONTROLLERS_BASE_BINARY_DIR ${CONTROLLERS_BINARY_DIR}/base)
 
 set(CONTROLLERS_BASE_BINDING_DIR ${CONTROLLERS_BASE_DIR}/binding)
 set(CONTROLLERS_BASE_BINDING_INCLUDE_DIR ${CONTROLLERS_BASE_BINDING_DIR}/include)
