+++ morphac/planners/base/CMakeLists.txt
@@ -1,20 +1,20 @@
 # Directories
 # -------------------------------------------------
 
-set(PLANNERS_DIR ${CMAKE_CURRENT_SOURCE_DIR})
-set(PLANNERS_SRC_DIR ${PLANNERS_DIR}/src)
-set(PLANNERS_INCLUDE_DIR ${PLANNERS_DIR}/include)
-set(PLANNERS_TEST_DIR ${PLANNERS_DIR}/test)
-set(PLANNERS_PYTHON_DIR ${PLANNERS_DIR}/python)
+set(PLANNERS_BASE_DIR ${CMAKE_CURRENT_SOURCE_DIR})
+set(PLANNERS_BASE_SRC_DIR ${PLANNERS_BASE_DIR}/src)
+set(PLANNERS_BASE_INCLUDE_DIR ${PLANNERS_BASE_DIR}/include)
+set(PLANNERS_BASE_TEST_DIR ${PLANNERS_BASE_DIR}/test)
+set(PLANNERS_BASE_PYTHON_DIR ${PLANNERS_BASE_DIR}/python)
 
-set(PLANNERS_BINARY_DIR ${MORPHAC_BINARY_DIR}/planners)
+set(PLANNERS_BASE_BINARY_DIR ${PLANNERS_BINARY_DIR}/base)
 
-set(PLANNERS_BINDING_DIR ${PLANNERS_DIR}/binding)
-set(PLANNERS_BINDING_INCLUDE_DIR ${PLANNERS_BINDING_DIR}/include)
-set(PLANNERS_BINDING_SRC_DIR ${PLANNERS_BINDING_DIR}/src)
-set(PLANNERS_BINDING_BINARY_DIR ${PLANNERS_BINARY_DIR}/binding)
-set(PLANNERS_BINDING_INSTALL_DIR
-  ${PYTHON_BINDING_INSTALL_DIR}/planners
+set(PLANNERS_BASE_BINDING_DIR ${PLANNERS_BASE_DIR}/binding)
+set(PLANNERS_BASE_BINDING_INCLUDE_DIR ${PLANNERS_BASE_BINDING_DIR}/include)
+set(PLANNERS_BASE_BINDING_SRC_DIR ${PLANNERS_BASE_BINDING_DIR}/src)
+set(PLANNERS_BASE_BINDING_BINARY_DIR ${PLANNERS_BASE_BINARY_DIR}/binding)
+set(PLANNERS_BASE_BINDING_INSTALL_DIR
+  ${PLANNERS_BINDING_INSTALL_DIR}/base
 )
 
 # Adding subdirectories
@@ -27,12 +27,12 @@ add_subdirectory(binding)
 # -------------------------------------------------
 
 # Constructs source files.
-set(PLANNERS_SRC
+set(PLANNERS_BASE_SRC
 
   planner.cc
 )
 
-foreach(file ${PLANNERS_SRC})
+foreach(file ${PLANNERS_BASE_SRC})
   get_filename_component(lib_name
     ${file} NAME_WE
   )
@@ -41,13 +41,13 @@ foreach(file ${PLANNERS_SRC})
 
   add_library(${lib_name}
     SHARED
-    ${PLANNERS_SRC_DIR}/${file}
-    ${PLANNERS_INCLUDE_DIR}/${header_name}
+    ${PLANNERS_BASE_SRC_DIR}/${file}
+    ${PLANNERS_BASE_INCLUDE_DIR}/${header_name}
   )
 
   target_include_directories(${lib_name}
     PUBLIC
-    ${PLANNERS_INCLUDE_DIR}
+    ${PLANNERS_BASE_INCLUDE_DIR}
   )
 endforeach()
 
@@ -56,20 +56,20 @@ endforeach()
 # -------------------------------------------------
 
 # Simulation tests source files.
-set(PLANNERS_TEST_SRC
+set(PLANNERS_BASE_TEST_SRC
 
   planner_test.cc
 )
 
 # Creating the test executables.
-foreach(file ${PLANNERS_TEST_SRC})
+foreach(file ${PLANNERS_BASE_TEST_SRC})
   get_filename_component(test_name
     ${file} NAME_WE
   )
 
   add_executable(
     ${test_name}
-    ${PLANNERS_TEST_DIR}/${file}
+    ${PLANNERS_BASE_TEST_DIR}/${file}
   )
 
   add_test(NAME ${test_name}
