+++ morphac/planners/base/binding/include/planner_binding.h
@@ -4,18 +4,20 @@
 #include "pybind11/eigen.h"
 #include "pybind11/pybind11.h"
 
-#include "planners/include/planner.h"
+#include "planners/base/include/planner.h"
 
 namespace morphac {
 namespace planners {
+namespace base {
 namespace binding {
 
 // Trampoline class as the Planner class is abstract.
-class PyPlanner : public morphac::planners::Planner {};
+class PyPlanner : public morphac::planners::base::Planner {};
 
 void define_planner_binding(pybind11::module& m);
 
 }  // namespace binding
+}  // namespace base
 }  // namespace planners
 }  // namespace morphac
