+++ morphac/planners/base/binding/planners.cc
@@ -1,10 +1,11 @@
 #include "pybind11/eigen.h"
 #include "pybind11/pybind11.h"
 
-#include "planners/binding/include/planner_binding.h"
+#include "planners/base/binding/include/planner_binding.h"
 
 namespace morphac {
 namespace planners {
+namespace base {
 namespace binding {
 
 namespace py = pybind11;
@@ -14,5 +15,6 @@ PYBIND11_MODULE(_binding_planners_python, m) {
 }
 
 }  // namespace binding
+}  // namespace base
 }  // namespace planners
 }  // namespace morphac
