+++ morphac/planners/base/src/planner.cc
@@ -1,11 +1,13 @@
-#include "planners/include/planner.h"
+#include "planners/base/include/planner.h"
 
 namespace morphac {
 namespace planners {
+namespace base {
 
 Planner::Planner() {}
 
 Planner::~Planner() {}
 
+} // namespace base
 } // namespace planners
 } // namespace morphac
